refactor: merge layouts
