--- apps/main/src/loan/layout/index.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react'
-
-const BaseLayout = ({ children }: { children: ReactNode }) => children
-
-export default BaseLayout
